+++ Example/Helper.swift
@@ -26,6 +26,10 @@ struct Helper {
             return "Wait network to be established before doing this"
         case .waitLoadedMedia:
             return "Wait media to be loaded before doing this"
+        case .itemDurationNotSet:
+            return "Media duration needed to perform action"
+        case .seekOverstepTime:
+            return "You can't seek after item end"
         }
     }
 }